# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/prospects": {
      "post": {
        "operationId": "ProspectsController_create",
        "summary": "Create a new prospect",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProspectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The prospect has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "prospects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProspectsController_findAll",
        "summary": "Get all prospects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all prospects."
          },
          "404": {
            "description": "Prospects not found."
          }
        },
        "tags": [
          "prospects"
        ]
      }
    },
    "/prospects/{id}": {
      "get": {
        "operationId": "ProspectsController_findOne",
        "summary": "Get a prospect by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a prospect."
          },
          "404": {
            "description": "Prospect not found."
          }
        },
        "tags": [
          "prospects"
        ]
      },
      "put": {
        "operationId": "ProspectsController_update",
        "summary": "Update a prospect by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProspectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The prospect has been successfully updated."
          },
          "404": {
            "description": "Prospect not found."
          }
        },
        "tags": [
          "prospects"
        ]
      },
      "delete": {
        "operationId": "ProspectsController_remove",
        "summary": "Delete a prospect by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The prospect has been successfully deleted."
          },
          "404": {
            "description": "Prospect not found."
          }
        },
        "tags": [
          "prospects"
        ]
      }
    },
    "/prospects/me": {
      "get": {
        "operationId": "ProspectsController_findAllOwner",
        "summary": "Get all prospects for the current owner",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all prospects for the current owner."
          },
          "404": {
            "description": "Prospects not found."
          }
        },
        "tags": [
          "prospects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api ZenFlow",
    "description": "Api for a CRM/Forum application with modules.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CustomFieldValueDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProspectDto": {
        "type": "object",
        "properties": {
          "society": {
            "type": "string",
            "description": "The society of the prospect"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the prospect"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the prospect"
          },
          "email": {
            "type": "string",
            "description": "The email of the prospect"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the prospect"
          },
          "address": {
            "type": "string",
            "description": "The address of the prospect"
          },
          "status": {
            "type": "string",
            "description": "The status of the prospect"
          },
          "ownerId": {
            "type": "string",
            "description": "The owner ID of the prospect"
          },
          "lastContactDate": {
            "format": "date-time",
            "type": "string",
            "description": "The last contact date of the prospect"
          },
          "marketSegment": {
            "type": "string",
            "description": "The market segment of the prospect"
          },
          "needs": {
            "type": "string",
            "description": "The needs of the prospect"
          },
          "leadSource": {
            "type": "string",
            "description": "The lead source of the prospect"
          },
          "companySize": {
            "type": "string",
            "description": "The company size of the prospect"
          },
          "estimatedBudget": {
            "type": "number",
            "description": "The estimated budget of the prospect"
          },
          "customFieldValues": {
            "description": "The custom field values of the prospect",
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomFieldValueDto"
              }
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "address",
          "status",
          "ownerId"
        ]
      },
      "UpdateProspectDto": {
        "type": "object",
        "properties": {
          "society": {
            "type": "string",
            "description": "The society of the prospect"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the prospect"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the prospect"
          },
          "email": {
            "type": "string",
            "description": "The email of the prospect"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the prospect"
          },
          "address": {
            "type": "string",
            "description": "The address of the prospect"
          },
          "status": {
            "type": "string",
            "description": "The status of the prospect"
          },
          "ownerId": {
            "type": "string",
            "description": "The owner ID of the prospect"
          },
          "lastContactDate": {
            "format": "date-time",
            "type": "string",
            "description": "The last contact date of the prospect"
          },
          "marketSegment": {
            "type": "string",
            "description": "The market segment of the prospect"
          },
          "needs": {
            "type": "string",
            "description": "The needs of the prospect"
          },
          "leadSource": {
            "type": "string",
            "description": "The lead source of the prospect"
          },
          "companySize": {
            "type": "string",
            "description": "The company size of the prospect"
          },
          "estimatedBudget": {
            "type": "number",
            "description": "The estimated budget of the prospect"
          },
          "customFieldValues": {
            "description": "The custom field values of the prospect",
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomFieldValueDto"
              }
            ]
          }
        }
      },
...
    }
  }
}